[REFACTOR] refactor domain.credit
