+++ src/main/java/com/sullung2yo/seatcatcher/domain/credit/controller/CreditController.java
@@ -1,8 +1,8 @@
-package com.sullung2yo.seatcatcher.user.controller;
+package com.sullung2yo.seatcatcher.domain.credit.controller;
 
 import com.sullung2yo.seatcatcher.user.domain.User;
-import com.sullung2yo.seatcatcher.user.dto.request.CreditModificationRequest;
-import com.sullung2yo.seatcatcher.user.service.CreditService;
+import com.sullung2yo.seatcatcher.domain.credit.dto.request.CreditModificationRequest;
+import com.sullung2yo.seatcatcher.domain.credit.service.CreditService;
 import com.sullung2yo.seatcatcher.user.service.UserService;
 import io.swagger.v3.oas.annotations.Operation;
 import jakarta.validation.Valid;
