+++ src/main/java/com/sullung2yo/seatcatcher/domain/credit/dto/request/CreditModificationRequest.java
@@ -1,4 +1,4 @@
-package com.sullung2yo.seatcatcher.user.dto.request;
+package com.sullung2yo.seatcatcher.domain.credit.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Min;
