+++ src/main/java/com/sullung2yo/seatcatcher/domain/credit/enums/CreditPolicy.java
@@ -1,4 +1,4 @@
-package com.sullung2yo.seatcatcher.user.domain;
+package com.sullung2yo.seatcatcher.domain.credit.enums;
 
 
 import lombok.AllArgsConstructor;
