+++ src/main/java/com/sullung2yo/seatcatcher/domain/credit/service/CreditService.java
@@ -1,4 +1,4 @@
-package com.sullung2yo.seatcatcher.user.service;
+package com.sullung2yo.seatcatcher.domain.credit.service;
 
 import com.sullung2yo.seatcatcher.train.domain.YieldRequestType;
 import com.sullung2yo.seatcatcher.user.domain.User;
