+++ src/main/java/com/sullung2yo/seatcatcher/train/controller/UserTrainSeatController.java
@@ -9,8 +9,8 @@ import com.sullung2yo.seatcatcher.train.domain.YieldRequestType;
 // import com.sullung2yo.seatcatcher.train.dto.request.UserTrainSeatRequest;
 import com.sullung2yo.seatcatcher.train.service.SeatEventService;
 import com.sullung2yo.seatcatcher.train.service.UserTrainSeatService;
-import com.sullung2yo.seatcatcher.user.domain.CreditPolicy;
-import com.sullung2yo.seatcatcher.user.service.CreditService;
+import com.sullung2yo.seatcatcher.domain.credit.enums.CreditPolicy;
+import com.sullung2yo.seatcatcher.domain.credit.service.CreditService;
 import com.sullung2yo.seatcatcher.user.service.UserService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
