+++ src/test/java/com/sullung2yo/seatcatcher/user/controller/CreditControllerTest.java
@@ -6,7 +6,7 @@ import com.sullung2yo.seatcatcher.common.jwt.provider.JwtTokenProviderImpl;
 import com.sullung2yo.seatcatcher.user.domain.ProfileImageNum;
 import com.sullung2yo.seatcatcher.domain.auth.enums.Provider;
 import com.sullung2yo.seatcatcher.user.domain.User;
-import com.sullung2yo.seatcatcher.user.dto.request.CreditModificationRequest;
+import com.sullung2yo.seatcatcher.domain.credit.dto.request.CreditModificationRequest;
 import com.sullung2yo.seatcatcher.user.repository.UserRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
