+++ src/test/java/com/sullung2yo/seatcatcher/user/service/CreditServiceImplTest.java
@@ -1,6 +1,7 @@
 package com.sullung2yo.seatcatcher.user.service;
 
 import com.sullung2yo.seatcatcher.common.exception.UserException;
+import com.sullung2yo.seatcatcher.domain.credit.service.CreditServiceImpl;
 import com.sullung2yo.seatcatcher.train.domain.YieldRequestType;
 import com.sullung2yo.seatcatcher.user.repository.UserRepository;
 import com.sullung2yo.seatcatcher.user.domain.User;
